pyproject.toml
[project]
name = "composer"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "chromadb==0.6.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core==1.3.2",
    "joblib>=1.4.2",
    "jq>=1.8.0",
    "langchain-anthropic>=0.3.1",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.13",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.2.2",
    "langchain-openai>=0.2.14",
    "langchain-text-splitters>=0.3.4",
    "langgraph>=0.2.60",
    "lark>=1.2.2",
    "monty>=2025.1.9",
    "pathvalidate>=3.2.3",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "rapidocr-onnxruntime>=1.4.3",
    "requests-cache>=1.2.1",
    "xmltodict>=0.14.2",
]
dynamic = ["version"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "composer/_version.py"

[project.scripts]
composer = "composer.entrypoint:entrypoint"

[tool.ruff]
line-length = 120
exclude = ["composer/_version.py"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 60

[tool.uv.sources]
langchain-chroma = { git = "https://github.com/matthewcarbone/langchain", subdirectory = "libs/partners/chroma" }
